added four journal cards
